cambiar estado mascota

+++ src/main/java/ar/edu/davinci/adoptame/domain/Mascota.java
@@ -17,24 +17,24 @@ public class Mascota implements Serializable {
     private static final long serialVersionUID = 4293421164107278076L;
 
     @Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "id_mascota")
-	private Integer id;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id_mascota")
+    private Long id;
 
-	@Column(name="nombre", nullable = false, length = 45)
-	private String nombre;
+    @Column(name="nombre", nullable = false, length = 45)
+    private String nombre;
 
-	@Column(name = "sexo", nullable = false, length = 45)
-	private String sexo;
+    @Column(name = "sexo", nullable = false, length = 45)
+    private String sexo;
 
-	@Column(name = "tipo_mascota", length = 45)
-	private String tipoMascota;
+    @Column(name = "tipo_mascota", length = 45)
+    private String tipoMascota;
 
-	@Column(name="raza", nullable = false, length = 45)
-	private String raza;
+    @Column(name="raza", nullable = false, length = 45)
+    private String raza;
 
-	@Column(name="edad")
-	private Integer edad;
+    @Column(name="edad")
+    private Integer edad;
 
     @Column(name="foto_url")
     private String foto_url;
@@ -61,7 +61,7 @@ public class Mascota implements Serializable {
 
     public Mascota() {
 
-	}
+    }
 
     public Mascota(String nombre, String sexo, String tipoMascota, String raza, Integer edad, String foto_url, String estado, String descripcion, String tamanio, Usuario rescatista, Coordenada coordenadas) {
         this.nombre = nombre;
@@ -80,11 +80,11 @@ public class Mascota implements Serializable {
 
 
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
